Complete code splitting for classes between .hpp headers and .cpp implementations
