+++ direction.hpp
@@ -1,0 +1,3 @@
+#pragma once
+
+enum Direction {UP, RIGHT, DOWN, LEFT};
