+++ inavjaga.cpp
@@ -750,788 +750,6 @@ void deallocateAll() {
     }
 }
 
-Entity::Entity(char symbol, sista::Coordinates coordinates, ANSI::Settings& settings, Type type) :
-    sista::Pawn(symbol, coordinates, settings), type(type) {}
-
-
-Wall::Wall(sista::Coordinates coordinates, short int strength) :
-    Entity('#', coordinates, wallStyle, Type::WALL), strength(strength) {
-    Wall::walls.push_back(this);
-}
-void Wall::remove() {
-    Wall::walls.erase(std::find(Wall::walls.begin(), Wall::walls.end(), this));
-    field->erasePawn(this);
-    delete this;
-}
-bool Wall::getHit() {
-    if (--strength <= 0) {
-        // Verify if this creates a new breach by a DFS
-        std::stack<sista::Coordinates> dfs({
-            coordinates + directionMap[Direction::UP],
-            coordinates + directionMap[Direction::LEFT],
-            coordinates + directionMap[Direction::DOWN],
-            coordinates + directionMap[Direction::RIGHT]
-        });
-        std::set<sista::Coordinates> visited;
-        visited.insert({coordinates});
-        sista::Coordinates coords, breach;
-        bool foundBelowExit = false;
-        bool foundAboveExit = false;
-        while (!dfs.empty() && !foundBelowExit && !foundAboveExit) {
-            coords = dfs.top();
-            dfs.pop();
-
-            if (field->isOutOfBounds(coords)) continue; // Exiting the field
-            if (visited.count(coords)) continue; // Already visited
-            visited.insert(coords);
-
-            if (field->isOccupied(coords)) { // Cell is not free
-                Type type = ((Entity*)field->getPawn(coords))->type;
-                if (type == Type::WALL || type == Type::PORTAL) continue;
-            }
-            if (coords.y % (TUNNEL_UNIT * 3) == 0) { // Exiting the breach from below
-                foundBelowExit = true;
-                // The breach was right above for how neighbouring is defined in a grid
-                breach = coords + directionMap[Direction::UP];
-            }
-            if (coords.y % (TUNNEL_UNIT * 3) < TUNNEL_UNIT * 2) { // Exiting the breach on the upper side
-                foundAboveExit = true;
-                continue;
-            }
-            if (coords.x < TUNNEL_UNIT * 2
-                && (coords.y / TUNNEL_UNIT / 3) % 2 == 0) {
-                // On "even" tunnels we should not consider part of the breach the part to the left
-                continue;
-            } else if (coords.x >= WIDTH-(TUNNEL_UNIT * 2)
-                        && (coords.y / TUNNEL_UNIT / 3) % 2 == 1) {
-                // On "odd" tunnels we should not consider part of the breach the part to the right
-                continue;
-            }
-            if (foundBelowExit && foundAboveExit) break;
-
-            dfs.push(coords + directionMap[Direction::UP]);
-            dfs.push(coords + directionMap[Direction::LEFT]);
-            dfs.push(coords + directionMap[Direction::DOWN]);
-            dfs.push(coords + directionMap[Direction::RIGHT]);
-        }
-        if (foundBelowExit && foundAboveExit) {
-            if (breaches.count(breach.y)) {
-                breaches[breach.y].push_back(breach.x);
-            } else {
-                breaches[breach.y] = {breach.x};
-            }
-        }
-        this->remove();
-        return true;
-    }
-    return false;
-}
-std::bernoulli_distribution Wall::wearing(WALL_WEARING_PROBABILITY);
-ANSI::Settings Wall::wallStyle = {
-    RGB_ROCKS_FOREGROUND,
-    RGB_ROCKS_BACKGROUND,
-    ANSI::Attribute::BRIGHT
-};
-
-Bullet::Bullet(sista::Coordinates coordinates, Direction direction) :
-    Entity(directionSymbol[direction], coordinates, bulletStyle, Type::BULLET), direction(direction), collided(false) {
-    Bullet::bullets.push_back(this);
-}
-void Bullet::remove() {
-    Bullet::bullets.erase(std::find(Bullet::bullets.begin(), Bullet::bullets.end(), this));
-    field->erasePawn(this);
-    delete this;
-}
-void Bullet::move() {
-    sista::Coordinates next = this->coordinates + directionMap[direction];
-    if (field->isFree(next)) {
-        field->movePawn(this, next);
-    } else if (field->isOutOfBounds(next)) {
-        this->remove();
-    } else if (field->isOccupied(next)) {
-        Entity* entity = (Entity*)field->getPawn(next);
-        Type entityType = entity->type;
-        switch (entityType) {
-            case Type::WALL:
-                ((Wall*)entity)->getHit();
-                break;
-            case Type::MINE:
-                ((Mine*)entity)->trigger();
-                break;
-            case Type::ENEMY_BULLET: {
-                EnemyBullet* bullet = (EnemyBullet*)entity;
-                if (bullet->collided) return;
-                bullet->collided = true;
-                break;
-            }
-            case Type::BULLET: {
-                Bullet* bullet = (Bullet*)entity;
-                if (bullet->collided) return;
-                bullet->collided = true;
-                break;
-            }
-            case Type::ARCHER: {
-                ((Archer*)entity)->die();
-                break;
-            }
-            case Type::WORM_HEAD: {
-                Worm* worm = (Worm*)entity;
-                worm->getHit();
-                break;
-            }
-            case Type::WORM_BODY: {
-                WormBody* wormBody = (WormBody*)entity;
-                wormBody->die();
-                break;
-            }
-            default:
-                break;
-        }
-        this->remove(); // Hit something and the situation was not handled
-    }
-}
-ANSI::Settings Bullet::bulletStyle = {
-    ANSI::ForegroundColor::F_MAGENTA,
-    ANSI::BackgroundColor::B_BLACK,
-    ANSI::Attribute::BRIGHT
-};
-
-EnemyBullet::EnemyBullet(sista::Coordinates coordinates, Direction direction) :
-    Entity(directionSymbol[direction], coordinates, enemyBulletStyle, Type::ENEMY_BULLET), direction(direction), collided(false) {
-    EnemyBullet::enemyBullets.push_back(this);
-}
-void EnemyBullet::remove() {
-    EnemyBullet::enemyBullets.erase(std::find(EnemyBullet::enemyBullets.begin(), EnemyBullet::enemyBullets.end(), this));
-    field->erasePawn(this);
-    delete this;
-}
-void EnemyBullet::move() {
-    sista::Coordinates next = this->coordinates + directionMap[direction];
-    if (field->isFree(next)) {
-        field->movePawn(this, next);
-    } else if (field->isOutOfBounds(next)) {
-        this->remove();
-    } else if (field->isOccupied(next)) {
-        Entity* entity = (Entity*)field->getPawn(next);
-        Type entityType = entity->type;
-        switch (entityType) {
-            case Type::WALL:
-                ((Wall*)entity)->getHit();
-                break;
-            case Type::MINE:
-                ((Mine*)entity)->trigger();
-                break;
-            case Type::BULLET: {
-                Bullet* bullet = (Bullet*)entity;
-                if (bullet->collided) return;
-                bullet->collided = true;
-                break;
-            }
-            case Type::ENEMY_BULLET: {
-                EnemyBullet* bullet = (EnemyBullet*)entity;
-                if (bullet->collided) return;
-                bullet->collided = true;
-                break;
-            }
-            case Type::WORM_HEAD: {
-                Worm* worm = (Worm*)entity;
-                worm->getHit();
-                break;
-            }
-            case Type::WORM_BODY: {
-                WormBody* wormBody = (WormBody*)entity;
-                wormBody->die();
-                break;
-            }
-            case Type::PLAYER:
-                printEndInformation(EndReason::SHOT);
-                dead = true;
-                break;
-            default:
-                break;
-        }
-        this->remove(); // Hit something and the situation was not handled
-    }
-}
-ANSI::Settings EnemyBullet::enemyBulletStyle = {
-    ANSI::ForegroundColor::F_GREEN,
-    ANSI::BackgroundColor::B_BLACK,
-    ANSI::Attribute::BRIGHT
-};
-
-Chest::Chest(sista::Coordinates coordinates, Inventory inventory) : Entity('C', coordinates, chestStyle, Type::CHEST), inventory(inventory) {
-    Chest::chests.push_back(this);
-}
-void Chest::remove() {
-    Chest::chests.erase(std::find(Chest::chests.begin(), Chest::chests.end(), this));
-    field->erasePawn(this);
-    delete this;
-}
-ANSI::Settings Chest::chestStyle = {
-    ANSI::RGBColor(193, 201, 104),
-    RGB_BLACK,
-    ANSI::Attribute::REVERSE
-};
-
-Mine::Mine(sista::Coordinates coordinates) : Entity('*', coordinates, mineStyle, Type::MINE), triggered(false) {
-    Mine::mines.push_back(this);
-}
-void Mine::remove() {
-    Mine::mines.erase(std::find(Mine::mines.begin(), Mine::mines.end(), this));
-    field->erasePawn(this);
-    delete this;
-}
-void Mine::trigger() {
-    if (this->triggered) return;
-    this->triggered = true;
-    this->setSettings(Mine::triggeredMineStyle);
-    this->setSymbol('%');
-    field->rePrintPawn(this);
-}
-void Mine::explode() {
-    for (int j=-MINE_DAMAGE_RADIUS; j<=MINE_DAMAGE_RADIUS; j++) {
-        for (int i=-MINE_DAMAGE_RADIUS; i<=MINE_DAMAGE_RADIUS; i++) {
-            if (i == 0 && j == 0) continue;
-            sista::Coordinates target = this->coordinates + sista::Coordinates(j, i);
-            if (field->isOutOfBounds(target)) continue;
-            if (field->isOccupied(target)) {
-                Entity* entity = (Entity*)field->getPawn(target);
-                switch (entity->type) {
-                    case Type::MINE:
-                        ((Mine*)entity)->trigger();
-                        break;
-                    case Type::WALL: {
-                        int damage = mineDamage(rng);
-                        while (!((Wall*)entity)->getHit() && --damage);
-                        break;
-                    }
-                    case Type::PLAYER: case Type::PORTAL:
-                        break; // Player and Portals get no damage from mines
-                    case Type::ARCHER:
-                        ((Archer*)entity)->die();
-                        break;
-                    case Type::WORM_BODY:
-                        ((WormBody*)entity)->die();
-                        break;
-                    case Type::WORM_HEAD:
-                        ((Worm*)entity)->getHit();
-                        break;
-                    default:
-                        entity->remove();
-                }
-            }
-        }
-    }
-    this->remove();
-}
-std::bernoulli_distribution Mine::explosion(MINE_EXPLOSION_IN_FRAME_PROBABILITY);
-std::uniform_int_distribution<int> Mine::mineDamage(MINE_MINIMUM_DAMAGE, MINE_MAXIMUM_DAMAGE);
-ANSI::Settings Mine::mineStyle = {
-    ANSI::RGBColor(200, 100, 200),
-    RGB_ROCKS_FOREGROUND,
-    ANSI::Attribute::BRIGHT
-};
-ANSI::Settings Mine::triggeredMineStyle = {
-    RGB_BLACK,
-    ANSI::RGBColor(0xff, 0, 0),
-    ANSI::Attribute::BLINK
-};
-
-Archer::Archer(sista::Coordinates coordinates) : Entity('A', coordinates, archerStyle, Type::ARCHER) {
-    Archer::archers.push_back(this);
-}
-void Archer::move() {
-    // There is always a probability of a dumb move
-    if (dumbMoveDistribution(rng)) {
-        Direction direction = (Direction)(rand() % 4);
-        this->move(direction);
-        return;
-    }
-
-    auto [row, column] = coordinates;
-    sista::Coordinates above = coordinates + directionMap[Direction::UP];
-    // It could be right below a passage or breach
-    if (row % (TUNNEL_UNIT * 3) == 0) {
-        if (std::find(breaches[above.y].begin(), breaches[above.y].end(), column) != breaches[above.y].end()) {
-            // There is a breach right above, so try to step in
-            this->move(Direction::UP);
-            return;
-        } else if (std::find(passages[above.y].begin(), passages[above.y].end(), column) != passages[above.y].end()) {
-            // There is a passage, so try to ascend
-            this->move(Direction::UP);
-            return;
-        }
-    }
-    // It is in a passage or breach (in the proper range on y axis)
-    if (row % (TUNNEL_UNIT * 3) >= TUNNEL_UNIT * 2) {
-        // it uses BFS to pick the next move to get to the other side
-        std::queue<std::pair<sista::Coordinates, Direction>> bfs({
-            {coordinates + directionMap[Direction::UP], Direction::UP},
-            {coordinates + directionMap[Direction::LEFT], Direction::LEFT},
-            {coordinates + directionMap[Direction::DOWN], Direction::DOWN},
-            {coordinates + directionMap[Direction::RIGHT], Direction::RIGHT}
-        }); // {coordinate, first direction taken in the path}
-        std::set<sista::Coordinates> visited;
-        visited.insert({coordinates});
-        Direction chosenMove;
-        bool found = false;
-        while (!bfs.empty()) {
-            auto [coords, choice] = bfs.front();
-            bfs.pop();
-
-            if (field->isOutOfBounds(coords)) continue; // Exiting the field
-            if (std::find(visited.begin(), visited.end(), coords) != visited.end()) continue; // Already visited
-            if (field->isOccupied(coords)) { // Cell is not free
-                Type type = ((Entity*)field->getPawn(coords))->type;
-                if (type == Type::WALL || type == Type::PORTAL) continue;
-            }
-            if (coords.y % (TUNNEL_UNIT * 3) == 0) continue; // Exiting the breach
-            if (coords.y % (TUNNEL_UNIT * 3) < TUNNEL_UNIT * 2) { // Exiting the breach on the upper side
-                chosenMove = choice;
-                found = true;
-                break;
-            }
-            visited.insert(coords);
-
-            bfs.push({coords + directionMap[Direction::UP], choice});
-            bfs.push({coords + directionMap[Direction::LEFT], choice});
-            bfs.push({coords + directionMap[Direction::DOWN], choice});
-            bfs.push({coords + directionMap[Direction::RIGHT], choice});
-        }
-        if (found) {
-            this->move(chosenMove);
-            return;
-        } else {
-            Direction direction = (Direction)(rand() % 4);
-            this->move(direction);
-            return;
-        }
-    }
-    // It roughly (probabilistically by x/y ratio) points to the centermost (or a random-but-deterministic one) breach
-    int next_passage_y = (row - (row % (TUNNEL_UNIT * 3))) - 1;
-    int next_passage_x;
-    if (!breaches[next_passage_y].empty()) {
-        std::vector<int> centermost_breaches = breaches[next_passage_y];
-        if (!centermost_breaches.empty()) {
-            std::sort(centermost_breaches.begin(), centermost_breaches.end(), [](int a, int b) {
-                return std::abs(a - WIDTH/2) < std::abs(b - WIDTH/2);
-            });
-            next_passage_x = centermost_breaches[(reinterpret_cast<intptr_t>(this)) % centermost_breaches.size()];
-        }
-    } else {
-        // If no breaches are found, then it points to any passage of the following level (random and not necessarily deterministic)
-        std::vector<int> available_passages = passages[next_passage_y];
-        if (!available_passages.empty()) {
-            next_passage_x = available_passages[rand() % available_passages.size()];
-        } else {
-            // If there is no passage on the next, it means we're pointing to a negative coordinate
-            // It is dangerous as generating coordinates from that could lead to segmentation faults
-            // but since we only use it to decide the next move it's actually innocuous
-            next_passage_x = WIDTH - 1; // Point to the top right
-        }
-    }
-    int delta_y = next_passage_y - row;
-    int delta_x = next_passage_x - column;
-    float ratio = (std::min(std::abs(delta_y), std::abs(delta_x)) + 1)
-                / (std::max(std::abs(delta_y), std::abs(delta_x)) + 1);
-    std::bernoulli_distribution verticalNotHorizontalDistribution(ratio);
-    if (verticalNotHorizontalDistribution(rng)) {
-        this->move(Direction::UP);
-    } else {
-        if (delta_x >= 0) {
-            this->move(Direction::RIGHT);
-        } else {
-            this->move(Direction::LEFT);
-        }
-    }
-}
-bool Archer::move(Direction direction) {
-    sista::Coordinates next = this->coordinates + directionMap[direction];
-    if (field->isOutOfBounds(next)) {
-        return false;
-    } else if (field->isFree(next)) {
-        field->movePawn(this, next);
-        return true;
-    } else if (field->isOccupied(next)) {
-        Entity* entity = (Entity*)field->getPawn(next);
-        switch (entity->type) {
-            case Type::MINE:
-                ((Mine*)entity)->trigger();
-            default:
-                return false;
-        }
-    }
-    return false;
-}
-void Archer::shoot() {
-    if (dumbMoveDistribution(rng)) {
-        this->shoot((Direction)(rand() % 4));
-        return;
-    }
-    if (coordinates.x == Player::player->getCoordinates().x) {
-        // Roughly vertically aligned with the player
-        if (coordinates.y / (TUNNEL_UNIT * 3) != Player::player->getCoordinates().y / (TUNNEL_UNIT * 3)) {
-            // They was a wall between them so the archer cannot see
-            int next_breaches_y = (coordinates.y % (TUNNEL_UNIT * 3)) - 1;
-            if (!breaches[next_breaches_y].empty()) {
-                if (std::find(breaches[next_breaches_y].begin(), breaches[next_breaches_y].end(), coordinates.x) != breaches[next_breaches_y].end()) {
-                    // ...unless there is a breach right above that allows the archer to hear the player moving behind the wall
-                    this->shoot(Direction::UP);
-                    return;
-                }
-            }
-            this->shoot((Direction)(rand() % 4));
-            return;
-        }
-        if (coordinates.y < Player::player->getCoordinates().y) {
-            this->shoot(Direction::DOWN);
-        } else if (coordinates.y > Player::player->getCoordinates().y) {
-            this->shoot(Direction::UP);
-        }
-        // Exactly vertically aligned with the player
-        if (coordinates.x == Player::player->getCoordinates().x) {
-            this->move(Direction::DOWN); // Dodges incoming bullets
-        }
-    } else if (coordinates.y == Player::player->getCoordinates().y) {
-        // Roughly horizontally aligned with the player
-        if (coordinates.x < Player::player->getCoordinates().x) {
-            this->shoot(Direction::RIGHT);
-        } else if (coordinates.x > Player::player->getCoordinates().x) {
-            this->shoot(Direction::LEFT);
-        }
-        // Exactly horizontally aligned with the player
-        if (coordinates.y == Player::player->getCoordinates().y) {
-            this->move(Direction::LEFT); // Dodges incoming bullets
-        }
-    } else {
-        this->shoot((Direction)(rand() % 4));
-    }
-}
-bool Archer::shoot(Direction direction) {
-    sista::Coordinates target = this->coordinates + directionMap[direction];
-    if (field->isOutOfBounds(target)) {
-        return false;
-    } else if (field->isFree(target)) {
-        field->addPrintPawn(new EnemyBullet(target, direction));
-        return true;
-    } else if (field->isOccupied(target)) {
-        Entity* entity = (Entity*)field->getPawn(target);
-        switch (entity->type) {
-            case Type::WALL:
-                ((Wall*)entity)->getHit();
-                break;
-            case Type::PLAYER: // Counts as a dagger hit
-                printEndInformation(EndReason::STABBED);
-                dead = true;
-                break;
-            case Type::MINE:
-                ((Mine*)entity)->trigger();
-                break;
-            default:
-                break;
-        }
-        return false;
-    }
-    return false;
-}
-void Archer::die() {
-    Archer::archers.erase(std::find(Archer::archers.begin(), Archer::archers.end(), this));
-    field->erasePawn(this);
-    field->addPrintPawn(new Chest(coordinates, {
-        LOOT_ARCHER_CLAY,
-        LOOT_ARCHER_BULLETS,
-        LOOT_ARCHER_MEAT
-    }));
-    delete this;
-}
-void Archer::remove() {
-    Archer::archers.erase(std::find(Archer::archers.begin(), Archer::archers.end(), this));
-    field->erasePawn(this);
-    delete this;
-}
-std::bernoulli_distribution Archer::moving(ARCHER_MOVING_PROBABILITY);
-std::bernoulli_distribution Archer::shooting(ARCHER_SHOOTING_PROBABILITY);
-std::bernoulli_distribution Archer::spawning(ARCHER_SPAWNING_PROBABILITY);
-ANSI::Settings Archer::archerStyle = {
-    ANSI::ForegroundColor::F_CYAN,
-    ANSI::BackgroundColor::B_BLACK,
-    ANSI::Attribute::STRIKETHROUGH
-};
-
-WormBody::WormBody(sista::Coordinates coordinates, Direction direction) : Entity(directionSymbol[direction], coordinates, wormBodyStyle, Type::WORM_BODY) {
-    WormBody::wormBodies.push_back(this);
-}
-void WormBody::die() {
-    sista::Coordinates drop = this->coordinates;
-    #if DEBUG
-    std::cerr << "WormBody::die() called for " << this << " at {" << drop.y << ", " << drop.x << "}" << std::endl;
-    #endif
-    field->addPrintPawn(new Chest(drop, {0, 1, 0}));
-    WormBody::wormBodies.erase(std::find(WormBody::wormBodies.begin(), WormBody::wormBodies.end(), this));
-    this->head->body.erase(std::find(this->head->body.begin(), this->head->body.end(), this));
-    delete this;
-}
-void WormBody::remove() {
-    WormBody::wormBodies.erase(std::find(WormBody::wormBodies.begin(), WormBody::wormBodies.end(), this));
-    field->erasePawn(this);
-    delete this;
-}
-ANSI::Settings WormBody::wormBodyStyle = {
-    ANSI::RGBColor(50, 0xff, 150),
-    ANSI::BackgroundColor::B_BLACK,
-    ANSI::Attribute::BRIGHT
-};
-
-Worm::Worm(sista::Coordinates coordinates) : Entity('H', coordinates, wormHeadStyle, Type::WORM_HEAD), hp(WORM_HEALTH_POINTS), collided(false) {
-    direction = (Direction)(rand() % 4);
-    Worm::worms.push_back(this);
-}
-Worm::Worm(sista::Coordinates coordinates, Direction direction) : Worm(coordinates) {
-    this->direction = direction;
-}
-void Worm::move() {
-    sista::Coordinates oldHeadCoordinates = coordinates;
-    sista::Coordinates next = coordinates + directionMap[direction];
-    if (field->isOutOfBounds(next)) {
-        Direction toTheLeft = (Direction)((direction + 3) % 4);
-        Direction toTheRight = (Direction)((direction + 1) % 4);
-        Direction turningOptions[] = {toTheLeft, toTheRight};
-        Direction oldDirection = direction;
-        direction = turningOptions[rand() % 2];
-        next = coordinates + directionMap[direction];
-        if (field->isOutOfBounds(next)) {
-            if (direction == toTheLeft) {
-                direction = toTheRight;
-                next = coordinates + directionMap[direction];
-                if (field->isOutOfBounds(next)) {
-                    direction = oldDirection;
-                    this->getHit();
-                    return;
-                }
-            } else if (direction == toTheRight) {
-                direction = toTheLeft;
-                next = coordinates + directionMap[direction];
-                if (field->isOutOfBounds(next)) {
-                    direction = oldDirection;
-                    this->getHit();
-                    return;
-                }
-            }
-        }
-    }
-    if (field->isFree(next)) {
-        /* Movement inspired from Dune (https://github.com/Lioydiano/Dune/blob/90a1f9c412258f701e3dfe949b05c6bcaa171e9f/dune.cpp#L386) */
-        field->movePawn(this, next);
-        // We now create a piece of body to leave behind the head, a "neck"
-        WormBody* neck = new WormBody(oldHeadCoordinates, direction);
-        neck->head = this;
-        field->addPrintPawn(neck);
-        body.push_back(neck);
-        // Consider that we added a body piece, so we need to ensure it does not grow too much
-        if (body.size() > WORM_LENGTH) {
-            WormBody* tail = body.front();
-            sista::Coordinates drop = tail->getCoordinates();
-            #if DEBUG
-            std::cerr << "In Worm::move() deleting the tail piece " << this << " at {" << drop.y << ", " << drop.x << "}" << std::endl;
-            #endif
-            field->erasePawn(tail);
-            if (clayRelease(rng)) {
-                field->addPrintPawn(new Chest(drop, {1, 0, 0}));
-            }
-            body.erase(body.begin());
-            WormBody::wormBodies.erase(std::find(WormBody::wormBodies.begin(), WormBody::wormBodies.end(), tail));
-            delete tail;
-        }
-    } else if (field->isOccupied(next)) {
-        Entity* entity = (Entity*)field->getPawn(next);
-        switch (entity->type) {
-            case Type::PLAYER:
-                this->turn(options[rand() % 2]);
-                printEndInformation(EndReason::EATEN);
-                dead = true;
-                break;
-            case Type::WALL:
-                if (((Wall*)entity)->strength > 1)
-                    ((Wall*)entity)->getHit(); // They can weaken a wall but not destroy it
-                this->turn(options[rand() % 2]);
-                break;
-            case Type::WORM_HEAD:
-                if (((Worm*)entity)->hp <= 1) {
-                    ((Worm*)entity)->collided = true;
-                }
-                ((Worm*)entity)->getHit();
-            case Type::PORTAL:
-                this->turn(options[rand() % 2]);
-                break;
-            case Type::WORM_BODY:
-                if (!eatingTail(rng)) {
-                    this->turn(options[rand() % 2]);
-                    break;
-                }
-                ((WormBody*)entity)->die();
-                break;
-            case Type::ARCHER:
-                if (!eatingArcher(rng)) {
-                    this->turn(options[rand() % 2]);
-                    break;
-                }
-                break;
-            case Type::MINE:
-                ((Mine*)entity)->trigger();
-                this->turn(options[rand() % 2]);
-                break;
-            default:
-                entity->remove();
-        }
-    }
-}
-void Worm::turn() {
-    if (dumbMoveDistribution(rng)) {
-        this->turn(options[rand() % 2]);
-        return;
-    }
-    // TODO: proper turning intelligence
-    this->turn(options[rand() % 2]);
-}
-void Worm::turn(Direction direction_) {
-    if (direction_ == Direction::LEFT)
-        this->direction = (Direction)((this->direction + 3) % 4);
-    else if (direction_ == Direction::RIGHT)
-        this->direction = (Direction)((this->direction + 1) % 4);
-}
-void Worm::getHit() {
-    if (--hp <= 0) {
-        if (collided) return;
-        this->die();
-    }
-}
-void Worm::die() {
-    while (!body.empty()) {
-        WormBody* tail = body.front();
-        tail->die(); // Self deletes from the body too
-    }
-    Worm::worms.erase(std::find(Worm::worms.begin(), Worm::worms.end(), this));
-    field->erasePawn(this);
-    field->addPrintPawn(new Chest(coordinates, {
-        LOOT_WORM_HEAD_CLAY,
-        LOOT_WORM_HEAD_BULLETS,
-        LOOT_WORM_HEAD_MEAT
-    }));
-    delete this;
-}
-void Worm::remove() {
-    Worm::worms.erase(std::find(Worm::worms.begin(), Worm::worms.end(), this));
-    field->erasePawn(this);
-    delete this;
-}
-Direction Worm::options[2] = {Direction::LEFT, Direction::RIGHT};
-std::bernoulli_distribution Worm::turning(WORM_TURNING_PROBABILITY);
-std::bernoulli_distribution Worm::moving(WORM_MOVING_PROBABILITY);
-std::bernoulli_distribution Worm::spawning(WORM_SPAWNING_PROBABILITY);
-std::bernoulli_distribution Worm::eatingTail(WORM_EATING_TAIL_PROBABILITY);
-std::bernoulli_distribution Worm::eatingArcher(WORM_EATING_ARCHER_PROBABILITY);
-std::bernoulli_distribution Worm::clayRelease(CLAY_RELEASE_PROBABILITY);
-ANSI::Settings Worm::wormHeadStyle = {
-    ANSI::ForegroundColor::F_GREEN,
-    ANSI::BackgroundColor::B_BLACK,
-    ANSI::Attribute::RAPID_BLINK
-};
-
-
-Player::Player(sista::Coordinates coordinates) : Entity('$', coordinates, playerStyle, Type::PLAYER), mode(Player::Mode::COLLECT), inventory(INITIAL_INVENTORY) {}
-Player::Player() : Entity('$', {0, 0}, playerStyle, Type::PLAYER), mode(Player::Mode::COLLECT), inventory(INITIAL_INVENTORY) {}
-void Player::remove() {
-    field->erasePawn(this);
-    delete this;
-}
-void Player::move(Direction direction) {
-    sista::Coordinates next = this->coordinates + directionMap[direction];
-    if (field->isFree(next)) {
-        field->movePawn(this, next);
-    } else if (field->isOutOfBounds(next)) {
-        return;
-    } else if (field->isOccupied(next)) {
-        Entity* entity = (Entity*)field->getPawn(next);
-        switch (entity->type) {
-            case Type::PORTAL: {
-                Portal* portal = (Portal*)entity;
-                sista::Coordinates landing = portal->exit->getCoordinates() + directionMap[direction];
-                if (field->isFree(landing)) {
-                    field->movePawn(this, landing);
-                }
-                break;
-            }
-            default:
-                break;
-        }
-    }
-}
-void Player::shoot(Direction direction) {
-    sista::Coordinates target = this->coordinates + directionMap[direction];
-    if (!field->isFree(target)) {
-        if (field->isOutOfBounds(target)) return;
-        
-        Entity* entity = (Entity*)field->getPawn(target);
-        switch (this->mode) {
-            case Mode::COLLECT: {
-                if (entity->type == Type::CHEST) {
-                    Chest* chest = (Chest*)entity;
-                    this->inventory += chest->inventory;
-                    chest->remove();
-                }
-                break;
-            }
-            case Mode::DUMPCHEST: {
-                if (entity->type == Type::CHEST) {
-                    Chest* chest = (Chest*)entity;
-                    chest->inventory += {
-                        this->inventory.clay,
-                        this->inventory.bullets,
-                        0 // Meat cannot be deposited
-                    };
-                    this->inventory = {0, 0, this->inventory.meat};
-                }
-                break;
-            }
-            default:
-                return;
-        }
-        return;
-    }
-    
-    switch (this->mode) {
-        case Mode::BULLET:
-            if (--inventory.bullets >= 0) {
-                field->addPrintPawn(new Bullet(target, direction));
-            }
-            inventory.bullets = std::max(inventory.bullets, (short)0);
-            break;
-        case Mode::DUMPCHEST:
-            if (inventory.clay > 0 || inventory.bullets > 0) {
-                field->addPrintPawn(new Chest(target, {
-                    this->inventory.clay,
-                    this->inventory.bullets,
-                    0 // Meat cannot be deposited
-                }));
-                this->inventory = {0, 0, this->inventory.meat};
-            }
-            break;
-        case Mode::MINE:
-            if (this->inventory.containsAtLeast(COST_OF_MINE)) {
-                this->inventory -= COST_OF_MINE;
-                field->addPrintPawn(new Mine(target));
-            }
-            break;
-        default:
-            return;
-    }
-}
-ANSI::Settings Player::playerStyle = {
-    ANSI::ForegroundColor::F_RED,
-    ANSI::BackgroundColor::B_BLACK,
-    ANSI::Attribute::BRIGHT
-};
-
 std::unordered_map<Direction, sista::Coordinates> directionMap = {
     {Direction::UP, {(unsigned short)-1, 0}},
     {Direction::RIGHT, {0, 1}},
@@ -1561,7 +779,4 @@ void Inventory::operator-=(const Inventory& other) {
     clay -= std::max(other.clay, (short)0);
     bullets -= std::max(other.bullets, (short)0);
     meat -= std::max(other.meat, (short)0);
-}
-inline bool Inventory::containsAtLeast(const Inventory other) const {
-    return clay >= other.clay && bullets >= other.bullets && meat >= other.meat;
 }
