+++ inventory.hpp
@@ -1,3 +1,5 @@
+#pragma once
+
 struct Inventory {
     short clay;
     short bullets;
@@ -6,5 +8,7 @@ struct Inventory {
     void operator+=(const Inventory&);
     void operator-=(const Inventory&);
 
-    inline bool containsAtLeast(const Inventory) const;
+    inline bool containsAtLeast(const Inventory other) const {
+        return clay >= other.clay && bullets >= other.bullets && meat >= other.meat;
+    }
 }; // The idea is that the inventory can be dropped (as CHEST) and picked up by the player
